+++ uniteai/document/demo.py
@@ -1,5 +1,5 @@
-from uniteai.document_chat import download as d
-from uniteai.document_chat import embed as e
+from uniteai.document import download as d
+from uniteai.document import embed as e
 
 from InstructorEmbedding import INSTRUCTOR
 from dataclasses import dataclass
@@ -77,9 +77,6 @@ if False:
             loaded[url] = dl.fetch_utf8(url)
 
 
-##################################################
-#
-
 def load_from_database(connection, title: Union[str, None], url):
     cursor = connection.cursor()
     if title is not None:
@@ -96,74 +93,73 @@ WHERE url = ?
     return result if result else (None, None, None, None)
 
 
-try:
-    already_loaded
-except:
-    # model = INSTRUCTOR('hkunlp/instructor-base')
-
-    # model = SentenceTransformer('sentence-transformers/all-mpnet-base-v2')
-    model = SentenceTransformer('sentence-transformers/all-MiniLM-L6-v2')
-    already_loaded = True
-
-embedding = e.Embedding(model, EMBEDDING_CACHE)
-search = e.Search(model,
-                  EMBEDDING_CACHE,
-                  embedding,
-                  percentile=90,)
-
-
-def search_helper(connection,
-                  titles_urls,
-                  query,
-                  window_size=200, stride=50, top_n=5, visualize=False):
-    results = []
-    for title, url in titles_urls:
-        title, file_path, content, url = load_from_database(connection, title, url)
-        if not title:
-            continue
-        result = search.search(file_path, query, window_size, stride, top_n, visualize)
-        results.append((url, title, result))
-    return results
-
-
 ##################################################
-# Go
+# Test
 
-titles_urls = [
-    # ('alice in wonderland', 'https://www.gutenberg.org/cache/epub/11/pg11.txt'),
-    # (None, 'https://github.com/normal-computing/outlines'),
-    (None, 'https://www.youtube.com/watch?v=Se91Pn3xxSs'),
-]
+if False:
+    try:
+        already_loaded
+    except:
+        # model = INSTRUCTOR('hkunlp/instructor-base')
 
-# query = 'where is the scene where her size changes?'
-# query = 'the most psychedelic scene.'
-# query = 'sea creatures'
-# query = 'the girl meets a cat for the first time'
-# query = 'the girl meets a cat in a tree for the first time'
-# query = 'alice is asleep'
-# query = 'alice meets royalty'
-# query = 'alice talks to a king'
-# query = 'alice talks to a queen'
-# query = 'alice talks to a rabbit'
-# query = 'the chapter title'
+        # model = SentenceTransformer('sentence-transformers/all-mpnet-base-v2')
+        model = SentenceTransformer('sentence-transformers/all-MiniLM-L6-v2')
+        already_loaded = True
 
-# query = 'a benefit of artificial intelligence'
-# query = 'the greatest dangers from artificial intelligence'
-# query = 'an impact on artists'
-query = 'a new industry will come from A.I.'
+    embedding = e.Embedding(model, EMBEDDING_CACHE)
+    search = e.Search(model,
+                      EMBEDDING_CACHE,
+                      embedding,
+                      percentile=90,)
 
-with sqlite3.connect(DB_PATH) as conn:
-    xs = search_helper(conn, titles_urls, query,
-                       window_size=500,
-                       stride=50,
-                       top_n=5,
-                       visualize=True)
+    def search_helper(connection,
+                      titles_urls,
+                      query,
+                      window_size=200, stride=50, top_n=5, visualize=False):
+        results = []
+        for title, url in titles_urls:
+            title, file_path, content, url = load_from_database(connection, title, url)
+            if not title:
+                continue
+            result = search.search(file_path, query, window_size, stride, top_n, visualize)
+            results.append((url, title, result))
+        return results
 
-print('\n'*20)
-for url, title, res in xs:
-    print('%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%')
-    print('%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%')
-    print(f'TITLE: {title}')
-    for r in res:
-        print('------------------------------')
-        print(r)
+    titles_urls = [
+        # ('alice in wonderland', 'https://www.gutenberg.org/cache/epub/11/pg11.txt'),
+        # (None, 'https://github.com/normal-computing/outlines'),
+        (None, 'https://www.youtube.com/watch?v=Se91Pn3xxSs'),
+    ]
+
+    # query = 'where is the scene where her size changes?'
+    # query = 'the most psychedelic scene.'
+    # query = 'sea creatures'
+    # query = 'the girl meets a cat for the first time'
+    # query = 'the girl meets a cat in a tree for the first time'
+    # query = 'alice is asleep'
+    # query = 'alice meets royalty'
+    # query = 'alice talks to a king'
+    # query = 'alice talks to a queen'
+    # query = 'alice talks to a rabbit'
+    # query = 'the chapter title'
+
+    # query = 'a benefit of artificial intelligence'
+    # query = 'the greatest dangers from artificial intelligence'
+    # query = 'an impact on artists'
+    query = 'a new industry will come from A.I.'
+
+    with sqlite3.connect(DB_PATH) as conn:
+        xs = search_helper(conn, titles_urls, query,
+                           window_size=500,
+                           stride=50,
+                           top_n=5,
+                           visualize=True)
+
+    print('\n'*20)
+    for url, title, res in xs:
+        print('%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%')
+        print('%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%')
+        print(f'TITLE: {title}')
+        for r in res:
+            print('------------------------------')
+            print(r)
